Remove padding from arrowMiddleware
We don't really need this padding, especially with that big value.
With that big value we see how buggy arrow shifts when we are close
to the edge of the view port

Trigger tooltip position update when the element is hovered,
but not focused.
It is possible to loose the focus when we moving the Thumb,
but we were not remeasuring the position on just hover.